requirements.txt
streamlit
spacy
transformers
torch
torchvision
Pillow
